+++ src/Engine/Animation/AnimationController.cpp
@@ -27,7 +27,7 @@ void AnimationController::Save(const std::string& path) {
         return;
     }
 
-    uint32_t numNodes = animationNodes.size();
+    uint32_t numNodes = static_cast<uint32_t>(animationNodes.size());
     file.write(reinterpret_cast<char*>(&numNodes), sizeof(uint32_t));
 
     for (Node* node : animationNodes) {
@@ -48,13 +48,13 @@ void AnimationController::Save(const std::string& path) {
         }
     }
 
-    uint32_t numBools = boolMap.size();
+    uint32_t numBools = static_cast<uint32_t>(boolMap.size());
     file.write(reinterpret_cast<char*>(&numBools), sizeof(uint32_t));
 
     for (BoolItem* b : boolMap)
         file.write(reinterpret_cast<char*>(b), sizeof(BoolItem));
 
-    uint32_t numFloats = floatMap.size();
+    uint32_t numFloats = static_cast<uint32_t>(floatMap.size());
     file.write(reinterpret_cast<char*>(&numFloats), sizeof(uint32_t));
 
     for (FloatItem* f : floatMap)
